requirements.txt
agno
google-genai
python-dotenv
beautifulsoup4
requests
pydantic
